fix circ dep
